+++ change/@vonage-vivid-77e875dc-eed2-4ceb-b2ad-1c64c283f771.json
@@ -1,6 +1,6 @@
 {
   "type": "prerelease",
-  "comment": "add textfield component",
+  "comment": "add text field component",
   "packageName": "@vonage/vivid",
   "email": "yonatan.kra@vonage.com",
   "dependentChangeType": "patch"
