+++ libs/components/src/lib/text-field/index.ts
@@ -4,8 +4,8 @@ import '../focus';
 import type { FoundationElementDefinition } from '@microsoft/fast-foundation';
 import { designSystem } from '../../shared/design-system';
 import { TextField } from './text-field';
-import styles from './textfield.scss';
-import { TextfieldTemplate as template } from './textfield.template';
+import styles from './text-field.scss';
+import { TextfieldTemplate as template } from './text-field.template';
 
 export const vividTextfield = TextField.compose<FoundationElementDefinition>({
 	baseName: 'text-field',
