+++ libs/components/src/lib/text-field/text-field.scss
@@ -1,6 +1,6 @@
 @use '../focus/partials/variables' as focus-variables;
 
-// @use '../../../../shared/src/lib/sass/mixins/typography' as typography;
+@use '../../../../shared/src/lib/sass/mixins/typography' as typography;
 @use '../../../../../node_modules/@vonage/vivid-tokens/dist/sass/themes/consts' as theme-consts;
 
 @use '../../../../shared/src/lib/sass/mixins/connotation/config' with (
@@ -24,6 +24,12 @@
 	@include connotation.connotation;
 	@include appearance.appearance;
 
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.fieldset {
 	position: relative;
 	display: flex;
 	align-items: center;
@@ -32,11 +38,10 @@
 	box-shadow: inset 0 0 0 1px var(#{appearance.get-variable(outline)});
 	color: var(#{appearance.get-variable(text)});
 	contain: strict;
-	cursor: text;
+	// cursor: text;
 	padding-inline: 16px;
 	transition: color .2s, box-shadow .2s, background-color .2s;
 }
-
 // .control-wrapper {
 // 	position: relative;
 // 	display: flex;
@@ -45,6 +50,9 @@
 // }
 
 .control {
+	@include typography.typography-cat-shorthand('body-2');
+
+  // required to stretch fully so autofill indication applies on all fieldset region
 	position: absolute;
 
 	// to avoid autofill style hiding other elements
@@ -55,33 +63,47 @@
 	color: var(#{theme-consts.$vvd-color-on-canvas});
 	inset: 0;
 	outline: 0 none;
+
+  .base:not(.icon) & {
+    padding-inline-start: 16px;
+  }
+  .base.icon & {
+    padding-inline-start: 44px;
+  }
+
+  &::placeholder {
+    color: var(#{theme-consts.$vvd-color-neutral-70});
+  }
 }
 
 .label {
-	transform-origin: 0;
-	transition: transform 0.2s;
+	@include typography.typography-cat-shorthand('body-2');
+// 	transform-origin: 0;
+// 	transition: transform 0.2s;
 
-	.base:where(.active, :focus-within) & {
-		transform: translateY(-8px) scale(0.7);
-	}
+// 	.base:where(.active, :focus-within) & {
+// 		transform: translateY(-8px) scale(0.7);
+// 	}
 }
 
 /* Shape */
 
-.base:not(.shape-pill) {
-	border-radius: 6px;
-}
-.base.shape-pill {
-	border-radius: 24px;
+.fieldset {
+  .base:not(.shape-pill) & {
+    border-radius: 6px;
+  }
+  .base.shape-pill & {
+    border-radius: 24px;
+  }
 }
 
 .focus-indicator {
-	.base:not(:focus-within) > & {
+	.fieldset:not(:focus-within) > & {
 		display: none;
 	}
 
-	.appearance-outlined &,
-	.appearance-ghost & {
+	.base.appearance-fieldset &,
+	.base.appearance-ghost & {
 		#{focus-variables.$focus-stroke-gap-color}: transparent;
 	}
 }
